requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
mssql-django==1.7
pyodbc==5.3.0
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
